[8.0][event_training_track] New module to mange training tracks.
- Source: Refactorization of https://github.com/OCA/event/pull/14.
- Design: [Google Docs][1].
- Target: https://github.com/OCA/l10n-spain/pull/193.
- Dependencies: https://github.com/OCA/event/pull/21 (bots will fail until merged).

This module takes `event_training` as a base and combines it with
`website_event_track`.

The result is that you can choose the expected duration for each training
event type and product. Events based on that product will be expected to
fulfill that amount of hours of each type. You can also check if those
expectations are fulfilled with a built-in wizard.

[1]: https://docs.google.com/document/d/1WiOB1yAWTR2aROpM6O632QWXwgG4hLRUpTHU2Pv4yEQ