Raise the right exception to fix tests.
The following tests were failing:

- test_grade_max_below_min
- test_grade_min_above_max
- test_grade_pass_above_max
- test_grade_pass_below_min

The reason of the failure is that a wrong exception was raised. This module
has been extensively refactored from https://github.com/OCA/event/pull/14
and that bug has been introduced in the refactoring process.

This little patch fixes that and makes the user get the right error message.