[ADD] event_registration_cancel_reason
=============================================
Reasons for event registrations cancellations
=============================================

This module allows to add a predefined reason why a registration for an event
is cancelled.

Configuration
=============

Go to *Marketing > Configuration > Events > Registrations cancel reasons* for
setting the possible reasons you want to handle. You can also set if a reason
is only available for certain events or, if you leave it empty, for all.

Usage
=====

Go to *Marketing > Events > Events*, enter into the details or any of the
events, and press the red circle button in the right part of any of the
registrations. A new screen will appear asking you to enter the reason why the
registration is cancelled.

You will get also the same screen when pressing the button
*Cancel registration* from the details view of a registration in
*Marketing > Events > Registrations*. On this object, you can group by the
cancel reason in the list, calendar or graph view.