[8.0][event_sale_registration_partner_unique] No problems with event_sale
This module combines the power of event_sale and event_registration_partner_unique to remove conflicts.

This is WIP until https://github.com/OCA/event/pull/38 is merged.

In case you confirm a sale order that would create a duplicate registration, then additional attendants are added to the already-existing registration, unless you are selling a different ticket type, in which case the duplicat partner is allowed (but not duplicate partner + ticket combinations).