[website_event_excerpt_img] Better layout for website events.
This displays an image taken from the event description and an excerpt in the events page, making it more attractive.

Add event priority to display them as half or full row on website.

Simplify dramatically removing options. All or nothing now.

This rewrites completely the event index layout, but it was impossible to get the desired effect without that due to core design, so I simplified everything.