crm_lead_to_event_registration (#32)
* [ADD] crm_lead_to_event_registration

====================================
Register a lead directly in an event
====================================

This module allows to register a lead directly in an event when converting
this lead into a partner, so that you can save several steps.

Usage
=====

Go to any lead in *Sales > Sales > Leads*, and click on the button "Convert
to Opportunity".

In the resulting screen, if the conversion action is "Convert to opportunity"
and you have selected a related partner (a new or an existing one), you will
have a section called "Event registration".

In that section, you can select an event, and when you click on
"Create opportunity", the partner will be registered automatically in the
event.

* [IMP] crm_lead_to_event_registration: Track state changes

* [IMP] crm_lead_to_event_registration: Adapt to latest company changes