[9.0][website_event_filter_selector] Migrate.
Relicensed to LGPL, new JS API.

Add test. Fix errors found with it.

Go 100% coverage. Support online events.

[FIX][website_event_filter_selector] Fix Travis build

Builds failed on 2017-03-31 because the tour was expecting to see USA event starting in this month, but next day was another month.

Well, in short, this fix makes the tour work on any day.