[FIX] website_event_snippet_calendar: Display event timezone
Before this patch, the user was getting the Public User's TZ on the
calendar view.

After merged, only the event's timzeone will be used to display dates.

It is, however, affected by https://github.com/odoo/odoo/issues/27179,
which means that there is a possible mismatch between the date used
to search and the one used to display, most likely to be hit when
events have very different timzeones. Added to Known Issues.