[ADD] website_event_questions_by_ticket: Conditional Events Questions
This module allows to set that an event question is only asked for certain
tickets.

If the question is general, then having one of the allowed tickets will serve
for showing it. If the question is specific to each attendee, then the
evaluation will be done for each of the tickets.

Configuration
=============

1. Go to *Events > Events*
2. Create or select one event.
3. On the "Questions" page, create a new question or edit an existing one.
4. Fill the field "Limited to tickets" for restricting the question to that
   ticket. Leave it blank for making it available for all tickets.

Usage
=====

1. Go to the website.
2. Go to "Events" section and select one of them.
3. Select ticket quantity and click on "Register".
4. On the popup dialog, the questions linked to that ticket and those without
   restrictions will be shown.

Known issues / Roadmap
======================

* Add tests (tour) to the module.