[MOD] 12.0 Replace code using system lang by babel dates translations
The big issue with this kind of code is that if you have a user with a
certain locale being used, it will override the locale system wide and
the next request on the server may have the dates formatted with
strftime in a different language than expected...

The other issue is that some systems do not provide a set of actual
locales so on a server running in docker without manually modifying
a dockerfile to install a set of locales, it may fail to switch
the locale to a non existing one. It will just fail to compute the field.

Babel is already a dependency of odoo and it's a bit strange to not use it.

Get the locale the same way ir.qweb.fields do

This commit use the same method as ir.qweb.fields to get the locale
to use to translate a date. The previous code didn't took into account
a case where the lang is undefined and this case also use the same
default language as defined in odoo en_US.

As a result, the behaviour of event_session is more in line with
the current behaviour in odoo which is to default to en_US if locale
is undefined. With the global locale setting, the default locale
would be the system one so it may give different result or strange
behaviours or simply prevent correctly working without manual
intervention on the host.. While babel is already supported and used
by odoo.

Modified the test to actually test correctly the date formatted

As the default lang is en_US the format of a date is the following:

- month/day/year

And time format is on 12 hours so not 22:00 to 23:00 but 10:00 PM to 11:00 PM

With those changes, setting the locale to en_UK,en_CA will render the correct date using

- day/month/year and hours on 24h.

Added tests to confirm the translations work correctly on some locales

Check case when locale=None should default to en_US
Check that lang change is correctly taken into account.
Check that it works with languages returning unicode text (Russian)