Update the event_sale_session to allow optional sale.order.line
When using the website flow, the sale.order.line won't always
be ready at the moment of the registration process. It's wrong
to expect the sale.order.line to be defined on the registration.

Also if the registration already provide a session_id we don't want
to override it.

event_sale works by creating as sale.order.line with the correct
config. Then the config creates the registration.

The website flow creates a registration then... creates a
a sale.order.line that will be affected to the registration.