[ADD] crm_event: combine leads and event types
This module extends the functionality of CRM opportunities (and leads, if
enabled) to support linking them to event types and to allow you to keep track
of leads interested in an upcoming event of some type.

This is useful if you organize your events based on the amount of people
interested in a certain type of event.

@Tecnativa TT27664