[FIX] website_event_crm: another company's event
The company field isn't mandatory for event types, so we have to check
that a lead auto invites to events available in the company of the lead.

We also have to check the proper base_url for our template. When we have
multi-website we want our customer going to the right link, otherwise
they'd be getting confusing info.

TT36027