[MIG] crm_event: Migration to version 17.0
This migration removes the `read_group` for event totals calculation in
the _compute_event_total method because the `seats_available` field in
v17 is not stored (`compute` and `store=False`). This makes `read_group`
unable to correctly process and calculate the value of the
`seats_available` field, as it operates directly on data stored
in the database.
Instead, a manual calculation is implemented which:
- Iterates over the filtered events according to the domain.
- Groups the data by `event_type_id`.
- Calculates the total number of available seats and the number of
events associated with each type.

TT51918