[FIX] website_event_filter_city: Country filter support
When the country filter is inactive, the node used is a span but when
the country filter is activated with the editor, the node is replaced
by a div and this causes an error. To avoid this, the node is replaced
by something generic ‘*’ and so it will be located by its identifying
class regardless of whether the element is a span or a div.